refactor: lazyFrame 적용
